game server commit

+++ world/physics.go
@@ -5,13 +5,13 @@ import (
 )
 
 type Position struct {
-	X float64
-	Y float64
+	X int64
+	Y int64
 }
 
 type Size struct {
-	Width  float64
-	Height float64
+	Width  int
+	Height int
 }
 
 type Physics struct {
@@ -27,26 +27,26 @@ type Vector struct {
 func GetDistance(p1 *Position, p2 *Position) float64 {
 	dltx := p2.X - p1.X
 	dlty := p2.Y - p1.Y
-	return math.Sqrt((dlty * dlty) + (dltx * dltx))
+	return math.Sqrt(float64(dlty*dlty) + float64(dltx*dltx))
 }
 
 func (p *Position) Add(v *Vector) Position {
 	return Position{
-		X: p.X + (math.Sin(v.Radians) * v.Distance),
-		Y: p.Y + (math.Cos(v.Radians) * v.Distance),
+		X: p.X + int64(math.Sin(v.Radians)*v.Distance),
+		Y: p.Y + int64(math.Cos(v.Radians)*v.Distance),
 	}
 }
 
 func (phys1 *Physics) CheckCollision(phys2 *Physics) bool {
 	p1l := phys1.X
-	p1r := p1l + phys1.Width
+	p1r := p1l + int64(phys1.Width)
 	p1t := phys1.Y
-	p1b := p1t + phys1.Height
+	p1b := p1t + int64(phys1.Height)
 
 	p2l := phys2.X
-	p2r := p2l + phys2.Width
+	p2r := p2l + int64(phys2.Width)
 	p2t := phys2.Y
-	p2b := p2t + phys2.Height
+	p2b := p2t + int64(phys2.Height)
 
 	return p1l < p2r && p1r > p2l && p1t < p2b && p1b > p2t
 }
